mvvm beépítése, két fragment a tevékenység röggzítéséhez, gyakorlat választó és maga az edzés felvétele
