+++ app/src/main/res/layout/mvvm_activity_test.xml
@@ -1,80 +1,53 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
 
     <data>
+
     </data>
 
     <androidx.coordinatorlayout.widget.CoordinatorLayout
-
-
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        tools:context=".mvvm.ui.fortabs.MvvmGyakorlatValasztoFragment">
 
         <com.google.android.material.appbar.AppBarLayout
-            android:id="@+id/appbar"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_width="match_parent">
+            android:theme="@style/AppTheme.AppBarOverlay">
+
+            <com.google.android.material.appbar.CollapsingToolbarLayout
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:fitsSystemWindows="true"
+                app:contentScrim="#000000"
+                app:titleEnabled="false"
+                app:layout_scrollFlags="scroll|enterAlways|exitUntilCollapsed"
+                app:toolbarId="@id/toolbar">
 
                 <androidx.appcompat.widget.Toolbar
-                    android:id="@+id/toolbar"
-                    android:layout_height="?attr/actionBarSize"
                     android:layout_width="match_parent"
-                    app:layout_collapseMode="pin"
-                    android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar" />
+                    android:layout_height="?attr/actionBarSize"
+                    android:id="@+id/toolbar"
+                    android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
+                    app:layout_collapseMode="pin" />
+
+                <com.google.android.material.tabs.TabLayout
+                    android:id="@+id/tabs"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:background="?attr/colorPrimary"
+                    android:layout_marginTop="?attr/actionBarSize"
+                    app:layout_collapseMode="pin" />
+
+            </com.google.android.material.appbar.CollapsingToolbarLayout>
         </com.google.android.material.appbar.AppBarLayout>
 
-          <LinearLayout
-              android:layout_width="match_parent"
-              android:layout_height="match_parent"
-              android:orientation="vertical"
-              android:layout_marginTop="?attr/actionBarSize">
-              <LinearLayout
-                  android:layout_width="match_parent"
-                  android:layout_height="wrap_content"
-                  android:orientation="vertical"
-                  android:background="#0089C8">
-                  <TextView
-                      android:id="@+id/textView16"
-                      android:layout_width="match_parent"
-                      android:layout_height="wrap_content"
-                      android:text="Gyakorlatok"
-                      android:textAlignment="center"
-                      android:textSize="21sp"
-                      android:textStyle="bold" />
-                  <LinearLayout
-                      android:id="@+id/searcglayout"
-                      android:layout_width="match_parent"
-                      android:layout_height="wrap_content"
-                      android:orientation="horizontal"
-                      app:layout_constraintEnd_toEndOf="parent"
-                      app:layout_constraintStart_toStartOf="parent"
-                      app:layout_constraintTop_toBottomOf="@+id/tvTestNev">
-
-                      <SearchView
-                          android:id="@+id/gyak_search_view"
-                          android:layout_width="wrap_content"
-                          android:layout_height="match_parent"
-                          android:layout_weight="3"
-                          android:queryHint="Név vagy izomcsoport keresése..."/>
-                      <Spinner
-                          android:id="@+id/spinIzomcsop"
-                          android:layout_width="wrap_content"
-                          android:layout_height="match_parent"
-                          android:layout_weight="1"
-                          android:padding="3dp"
-                          android:spinnerMode="dialog" />
-                  </LinearLayout>
-              </LinearLayout>
-
-              <ListView
-                  android:id="@+id/lvGyakorlat"
-                  android:layout_width="match_parent"
-                  android:layout_height="match_parent"
-                  android:choiceMode="singleChoice"/>
-          </LinearLayout>
-
-
+        <androidx.viewpager.widget.ViewPager
+            android:id="@+id/view_pager"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            app:layout_behavior="@string/appbar_scrolling_view_behavior" />
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
 </layout>
